requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
cassandra-driver==3.28.0
cassio==0.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.2
datasets==2.15.0
dill==0.3.7
distro==1.8.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
geomet==0.2.1.post1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.340
langsmith==0.0.66
marshmallow==3.20.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.26.2
openai==0.27.0
packaging==23.2
pandas==2.1.3
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.3
